Add Collection.plot() function
As a first pass, I allow Collections to contain matplotlib.figure.Figure
objects. Collection.plot() requires a variable argument (for now a string, but
perhaps a list of variables to plot would be better?) and also allows the user
to pass isel_dict to help reduce dimensions. There is a trello card about
introducing sel / isel functionality, and once that's done the isel_dict
argument can be removed.

I like the idea of being able to super-impose line plots (time series of scalar
quantities, or vertical profiles) in a single figure, but I have not made any
progress on that front yet.

I also should add a test for this function.